# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Vulnerability-Info-API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Login Endpoint takes username and password and returns a JWT-Token if authorized.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Login Endpoint",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "Credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUpdateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.LoginSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "SingUp Endpoint takes username and password and returns a user object.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Authorization"
                ],
                "summary": "Sign Up Endpoint",
                "parameters": [
                    {
                        "description": "username and password",
                        "name": "Credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUpdateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/config/match": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Checks the components against vulnerabilities for a given time period.\nE.g.: '.../match?from=0?to=-10' covers Notifications created between today and 10 days ago.",
                "tags": [
                    "Config"
                ],
                "summary": "Check components for vulnerabilities",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "From days back: '0' = today. Must be \u003c 1.",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "To days back. Must be \u003c 0.",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/config/rss": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches the RSS Feed and runs the Parser, if new entries are available.",
                "tags": [
                    "Config"
                ],
                "summary": "Fetch RSS",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all listed notifications.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Get All notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Get notifications for a DFN-Link.",
                        "name": "link",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications for a specific CVSS base score.",
                        "name": "cvss_base",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications for a specific CVSS exploitability score.",
                        "name": "cvss_exploitability",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications for a specific CVSS impact score.",
                        "name": "cvss_impact",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications for a specific CVSS temporal score.",
                        "name": "cvss_temp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications associated with a specific CVE-ID. Will not succeed in combination with 'cve_id' or 'for'.",
                        "name": "cve_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Get all Notifications associated with or without a CVEs. Will not succeed in combination with 'cve' or 'for'.",
                        "name": "cve",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Get all Notifications that match components of a username (e-mail). Will not succeed in combination with 'cve' or 'cve_id'.",
                        "name": "for",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Notification"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Search notifications by title.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Search notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search notifications by title.",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Notification"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/notifications/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns notification for a specific ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Notification"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.ApiError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.LoginSuccess": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "helper.ApiError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateUpdateUserInput": {
            "type": "object",
            "required": [
                "e_mail",
                "password"
            ],
            "properties": {
                "e_mail": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.Notification": {
            "type": "object",
            "properties": {
                "cvss_base": {
                    "type": "string"
                },
                "cvss_exploitability": {
                    "type": "string"
                },
                "cvss_impact": {
                    "type": "string"
                },
                "cvss_temp": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "pub_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "vulnerabilities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Vulnerability"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "e_mail"
            ],
            "properties": {
                "e_mail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Vulnerability": {
            "type": "object",
            "properties": {
                "cve": {
                    "type": "string"
                },
                "cwe": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Use this token format: 'Bearer {key}'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}